pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph-rag-experiment"
version = "1.0.0"
dependencies = [
    "openai>=1.35.3",
    "sqlalchemy>=2.0.31",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.111.0",
    "pydantic>=2.7.4",
    "requests>=2.32.3",
    "torch>=2.3.1",
    "tqdm>=4.66.4",
    "transformers>=4.41.2",
    "uvicorn>=0.30.1",
    "lxml>=5.2.2",
    "pdf2image>=1.17.0",
    "pandas>=2.2.2",
    "anthropic>=0.29.0",
    "types-requests>=2.32.0.20240622",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.134.0",
    "uuid7>=0.1.0",
    "zipfile36>=0.1.3",
    "types-beautifulsoup4>=4.12.0.20240511",
    "pycryptodome>=3.20.0",
    "dataclasses-json>=0.6.7",
    "case-converter>=1.1.0",
    "pypdf>=4.2.0",
    "pymupdf>=1.24.7",
    "boto3>=1.34.143",
    "langchain>=0.2.12",
    "langchain-core>=0.2.28",
    "langchain-community>=0.2.11",
    "langchain-experimental>=0.0.64",
    "langchain-openai>=0.1.20",
    "neo4j>=5.23.1",
]

[tool.rye.scripts]
fmt = "rye run ruff format ./**/*.py"


[tool.rye]
dev-dependencies = [
    "google-api-python-client-stubs>=1.26.0",
    "mypy>=1.10.1",
    "google-auth-stubs>=0.2.0",
    "pandas-stubs>=2.2.2.240603",
    "debugpy>=1.8.2",
    "ruff>=0.5.0",
    "types-python-dateutil>=2.9.0.20240316",
    "boto3-stubs>=1.34.143",
    "mypy-boto3-s3>=1.34.138",
    "types-tqdm>=4.66.0.20240417",
]
requires-python = "~=3.11.9"
